+++ engine/qclib/progslib.h
@@ -132,7 +132,7 @@ struct progfuncs_s {
 
 	int lastcalledbuiltinnumber;			//useful with non-implemented opcodes.
 
-	int (*RegisterFieldVar)				(progfuncs_t *prinst, unsigned int type, char *name, int requestedpos, int originalofs);
+	int (*RegisterFieldVar)				(progfuncs_t *prinst, unsigned int type, char *name, size_t requestedpos, size_t originalofs);
 
 	char	*tempstringbase;				//for engine's use. Store your base tempstring pointer here.
 	int		tempstringnum;			//for engine's use.
